Added new matrix-mul varient for tutorial 3 and tutorial 4 for in-memory GEMM

+++ tutorial-utils/util-matrix.h
@@ -10,8 +10,8 @@
 #include <fstream>
 #include "enums/MatrixType.h"
 
-double *allocMatrix(int width, int height);
-void initMatrix(double *matrix, int width, int height);
+double *allocMatrix(long width, long height);
+void initMatrix(double *matrix, long width, long height);
 std::string generateFilename(std::string directory, MatrixType type, int blockRow, int blockCol, std::string suffix);
 std::string generateDirectoryName(std::string basePath, int totalWidth, int totalHeight, int blockSize);
 std::string generateDirectoryName(std::string basePath, int totalWidth, int totalHeight);
